Add sub-package for devtools
